Refactor fee_percentage type from u64 to u8 in coinflip program
- Changed the type of `fee_percentage` from `u64` to `u8` across multiple files, including configuration, state, and instruction handlers.
- Updated related calculations to ensure compatibility with the new type.
- Adjusted space allocation in account structs to reflect the change in data type.

These changes improve memory efficiency and align the fee percentage representation with its intended range.

+++ programs/coinflip/src/instructions/claim_winnings.rs
@@ -10,7 +10,7 @@ pub fn handler(ctx: Context<super::ClaimWinnings>) -> Result<()> {
         GameType::Coinflip => game.amount * (game.participants.len() as u64),
         GameType::Giveaway => game.amount,
     };
-    let fee_amount = total_pot * ctx.accounts.config.fee_percentage / 100;
+    let fee_amount = total_pot * (ctx.accounts.config.fee_percentage as u64) / 100;
     let winner_amount = total_pot - fee_amount;
 
     game.status = GameStatus::Completed;
